publish-component.sh: make remote publishing an explicit action
It may be desirable to create the subtree & tag locally and validate or
inspect results before pushing. Or it may be desirable to not push at
all. Furthermore, since tags cannot be namespaced across remotes the way
that branches can, pushing tags to remotes is somewhat of a big deal. In
light of all that, it seems wisest to only push the tag when a remote is
explicitly specified, and painstakingly avoid any magic around making
that decision.