publish-component.sh: simplify changelog generation
Finding the root commit when there hasn't been a release yet is not only
a needlessly complex action, but the changelog will then exclude the
root commit itself. Shortlog is smart enough to figure out what to do if
there is not a previous release to base our changelog on.