cicd.mk: add develop stage to ci orchestration
The design principals of stages are such that tests should not be
expected to pass before a stage itself has run. Also, the ordering of
stages in the ci/cd pattern is such that the integrate stage should not
be expected to run successfully unless the develop stage has done so.