gcp-secrets.mk: fix secret exists check in gcp-secrets-create
By using "secrets version describe", an error will result if a secret
exists but no versions have been created. Changing the command to
"secrets describe" still has the desired effect when a secret does not
exist (creates it), but does not cause an error when a secret exists but
has no versions.