engine: path-relto.sh: rename from dir-relto.sh
This re-assumes the old name, but the usage reflects more refined
semantics in which DEST can be any arbitrary path and START is expected
to be a directory. START must be a directory in any relative path
resolution tool such as this one; e.g, see the --relative-to flag in GNU
realpath.